ALL: Huge updated, see submessage
Huge update from Manjaro to Artix, all scripts are remade, cleaned
everything with dash posix-complaint scripts and changed to suckless
utilites.

I've been working on this for 2 months and I think it's finally time
to say it's ready, I can still improve the documentation and dotfiles
optimization in the future, but everything right now is good enough.

You can see all the applications and config files for each one are
included in the submodules, such as dwm, dmenu, st, lemonbar and slock.

I love Artix Linux running Runit and probably this is what I'll stop for
a long, long time.

Manjaro was a good run, but this, this is beautiful.